refactor(tui): extract copy overlay
